% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in-par.R
\name{parallel_method_get}
\alias{parallel_method_get}
\alias{parallel_method_set}
\title{Set parallel backend}
\usage{
parallel_method_get()

parallel_method_set(method = c("future", "mirai"))
}
\arguments{
\item{method}{method to use.}
}
\description{
Set parallel backend
}
